partie2

+++ tp4/purge.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 docker rm -f $(docker ps -aq)
+if [ $? != 0 ]; then
+	exit 2
+fi
 
 #purge  tous les réseaux
-docker volume prune -f
-
+docker network prune -f
+if [ $? != 0 ]; then
+	exit 2
+fi
 #purge  tous les volumes
 docker volume prune -f
+if [ $? != 0 ]; then
+	exit 2
+fi
+
